Fix docker-compose
