+++ Dockerfile
@@ -8,5 +8,9 @@ FROM alpine
 COPY --from=builder /build/main /app/
 WORKDIR /app 
 
+ENV WAIT_VERSION 2.7.2
+ADD https://github.com/ufoscout/docker-compose-wait/releases/download/$WAIT_VERSION/wait /wait
+RUN chmod +x /wait
+
 EXPOSE 8080/tcp
 CMD ["./main"]
