--- src/api/config/bbdd.go
@@ -1,1 +1,0 @@
-package config
