+++ src/go.mod
@@ -6,6 +6,6 @@ require (
 	github.com/gin-gonic/gin v1.6.3
 	golang.org/x/lint v0.0.0-20200302205851-738671d3881b // indirect
 	golang.org/x/tools v0.0.0-20201105220310-78b158585360 // indirect
-	gorm.io/driver/mysql v1.0.3 // indirect
-	gorm.io/gorm v1.20.6 // indirect
+	gorm.io/driver/mysql v1.0.3
+	gorm.io/gorm v1.20.6
 )
