+++ src/main.go
@@ -2,46 +2,18 @@ package main
 
 import (
 	server "github.com/fullstacktf/Narrativas-Backend/api/server"
-
-	"fmt"
-	"log"
-
-	"gorm.io/driver/mysql"
-	"gorm.io/gorm"
+	config "github.com/fullstacktf/Narrativas-Backend/config"
 )
 
-func DbURL(dbConfig *DBConfig) string {
-	return fmt.Sprintf(
-		"%s:%s@tcp(%s:%d)/%s?charset=utf8mb4&parseTime=True&loc=Local",
-		dbConfig.User,
-		dbConfig.Password,
-		dbConfig.Host,
-		dbConfig.Port,
-		dbConfig.DBName,
-	)
-}
-
-type DBConfig struct {
-	Host     string
-	Port     int
-	User     string
-	DBName   string
-	Password string
-}
-
 func main() {
-	config := &DBConfig{
-		Host:     "172.26.0.3",
+	dbdata := config.DBConfig{
+		Host:     "172.16.238.3",
 		Port:     3306,
 		User:     "rollify",
 		DBName:   "rollify",
 		Password: "password",
 	}
 
-	_, err := gorm.Open(mysql.Open(DbURL(config)), &gorm.Config{})
-	if err != nil {
-		log.Fatal("Error BBDD")
-	} else {
-		server.Serve()
-	}
+	db := config.DbConnect(dbdata)
+	server.Serve()
 }
